+++ Server/src/KeyLogger.java
@@ -1,4 +1,4 @@
-package keystroke;
+package src;
 
 import com.github.kwhat.jnativehook.keyboard.NativeKeyEvent;
 import com.github.kwhat.jnativehook.keyboard.NativeKeyListener;
