+++ Server/src/Server.java
@@ -1,4 +1,4 @@
-package keystroke;
+package src;
 
 import java.net.InetAddress;
 import java.net.ServerSocket;
